+++ analysis/src/CodeActions.ml
@@ -5,14 +5,18 @@ let stringifyCodeActions codeActions =
   Printf.sprintf {|%s|}
     (codeActions |> List.map Protocol.stringifyCodeAction |> Protocol.array)
 
-let make ~title ~kind ~uri ~newText ~range =
-  let uri = uri |> Uri.fromPath |> Uri.toString in
-  {
-    Protocol.title;
-    codeActionKind = kind;
-    edit =
-      {
-        documentChanges =
-          [{textDocument = {version = None; uri}; edits = [{newText; range}]}];
-      };
-  }
+let make ~title ~kind ~edit ~command =
+  {Protocol.title; codeActionKind = kind; edit; command}
+
+let makeEdit edits uri =
+  Protocol.
+    {
+      documentChanges =
+        [
+          {
+            textDocument =
+              {version = None; uri = uri |> Uri.fromPath |> Uri.toString};
+            edits;
+          };
+        ];
+    }