+++ analysis/src/Protocol.ml
@@ -3,7 +3,7 @@ type range = {start: position; end_: position}
 type markupContent = {kind: string; value: string}
 
 (* https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#command *)
-type command = {title: string; command: string}
+type command = {title: string; command: string; arguments: string list option}
 
 (* https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#codeLens *)
 type codeLens = {range: range; command: command option}
@@ -76,12 +76,13 @@ type textDocumentEdit = {
 }
 
 type codeActionEdit = {documentChanges: textDocumentEdit list}
-type codeActionKind = RefactorRewrite
+type codeActionKind = RefactorRewrite | Empty
 
 type codeAction = {
   title: string;
   codeActionKind: codeActionKind;
-  edit: codeActionEdit;
+  edit: codeActionEdit option;
+  command: command option;
 }
 
 let null = "null"
@@ -224,7 +225,7 @@ let stringifyoptionalVersionedTextDocumentIdentifier td =
     | Some v -> string_of_int v)
     (Json.escape td.uri)
 
-let stringifyTextDocumentEdit tde =
+let stringifyTextDocumentEdit (tde : textDocumentEdit) =
   Printf.sprintf {|{
   "textDocument": %s,
   "edits": %s
@@ -235,15 +236,37 @@ let stringifyTextDocumentEdit tde =
 let codeActionKindToString kind =
   match kind with
   | RefactorRewrite -> "refactor.rewrite"
+  | Empty -> ""
 
 let stringifyCodeActionEdit cae =
   Printf.sprintf {|{"documentChanges": %s}|}
     (cae.documentChanges |> List.map stringifyTextDocumentEdit |> array)
 
-let stringifyCodeAction ca =
-  Printf.sprintf {|{"title": "%s", "kind": "%s", "edit": %s}|} ca.title
-    (codeActionKindToString ca.codeActionKind)
-    (ca.edit |> stringifyCodeActionEdit)
+let stringifyCommand (command : command) =
+  stringifyObject
+    [
+      ("title", Some (wrapInQuotes command.title));
+      ("command", Some (wrapInQuotes command.command));
+      ( "arguments",
+        match command.arguments with
+        | None -> None
+        | Some args -> Some (args |> List.map wrapInQuotes |> array) );
+    ]
+
+let stringifyCodeAction (ca : codeAction) =
+  stringifyObject
+    [
+      ("title", Some (wrapInQuotes ca.title));
+      ("kind", Some (wrapInQuotes (codeActionKindToString ca.codeActionKind)));
+      ( "edit",
+        match ca.edit with
+        | None -> None
+        | Some edit -> Some (edit |> stringifyCodeActionEdit) );
+      ( "command",
+        match ca.command with
+        | None -> None
+        | Some command -> Some (command |> stringifyCommand) );
+    ]
 
 let stringifyHint hint =
   Printf.sprintf
@@ -256,12 +279,6 @@ let stringifyHint hint =
 }|}
     (stringifyPosition hint.position)
     (Json.escape hint.label) hint.kind hint.paddingLeft hint.paddingRight
-
-let stringifyCommand (command : command) =
-  Printf.sprintf {|{"title": "%s", "command": "%s"}|}
-    (Json.escape command.title)
-    (Json.escape command.command)
-
 let stringifyCodeLens (codeLens : codeLens) =
   Printf.sprintf
     {|{